backend/requirements.txt
# Core Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.27
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
bcrypt==4.1.2

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# AI & ML
openai==1.12.0
anthropic==0.18.1
langchain==0.1.7
langchain-community==0.0.20
sentence-transformers==2.3.1
qdrant-client==1.7.3

# Email
aiosmtplib==3.0.1
aioimaplib==1.0.1
email-validator==2.1.0
python-email==0.1.0

# Task Queue
celery==5.3.6
kombu==5.3.5

# Validation & Serialization
marshmallow==3.20.2
python-dateutil==2.8.2

# WebSocket
websockets==12.0
python-socketio==5.11.1

# Utilities
python-dotenv==1.0.1
pytz==2024.1
pyyaml==6.0.1
toml==0.10.2

# Monitoring & Logging
prometheus-client==0.19.0
sentry-sdk==1.40.0
structlog==24.1.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx-mock==0.12.1
faker==22.6.0

# Code Quality
ruff==0.2.1
mypy==1.8.0
black==24.2.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.9
